+++ Example/Tests/Tests/ZSWStringParserSpec.m
@@ -6,27 +6,48 @@
 //  Copyright (c) 2015 Zachary West. All rights reserved.
 //
 
-#import "ZSWStringParser.h"
-#import "ZSWStringParserTag.h"
+@import ZSWTaggedString.Private;
 
 SpecBegin(ZSWStringParser)
 
 describe(@"ZSWStringParser", ^{
     describe(@"when providing erroneous strings", ^{
-        it(@"should throw an exception if tags aren't ended", ^{
-            expect(^{
-                [ZSWStringParser stringWithTaggedString:[ZSWTaggedString stringWithString:@"<tag>moo"]
-                                                options:[ZSWTaggedStringOptions options]
-                                            returnClass:[NSString class]];
-            }).to.raise(NSInvalidArgumentException);
+        it(@"should return an error if tags aren't ended", ^{
+            NSError *error;
+            id string = [ZSWStringParser stringWithTaggedString:[ZSWTaggedString stringWithString:@"<tag>moo"]
+                                                        options:[ZSWTaggedStringOptions options]
+                                                    returnClass:[NSString class]
+                                                          error:&error];
+            
+            expect(string).to.beNil();
+            expect(error.domain).to.equal(ZSWTaggedStringErrorDomain);
+            expect(error.code).to.equal(ZSWTaggedStringErrorCodeInvalidTags);
+            
+            // and without an error arg
+            string = [ZSWStringParser stringWithTaggedString:[ZSWTaggedString stringWithString:@"<tag>moo"]
+                                                     options:[ZSWTaggedStringOptions options]
+                                                 returnClass:[NSString class]
+                                                       error:nil];
+            expect(string).to.beNil();
         });
         
         it(@"should throw an exception if tags are ended out-of-order", ^{
-            expect(^{
-                [ZSWStringParser stringWithTaggedString:[ZSWTaggedString stringWithString:@"<tag><elephant>moo</tag></elephant>"]
-                                                options:[ZSWTaggedStringOptions options]
-                                            returnClass:[NSString class]];
-            }).to.raise(NSInvalidArgumentException);
+            NSError *error;
+            
+            id string = [ZSWStringParser stringWithTaggedString:[ZSWTaggedString stringWithString:@"<tag><elephant>moo</tag></elephant>"]
+                                                        options:[ZSWTaggedStringOptions options]
+                                                    returnClass:[NSString class]
+                                                          error:&error];
+            expect(string).to.beNil();
+            expect(error.domain).to.equal(ZSWTaggedStringErrorDomain);
+            expect(error.code).to.equal(ZSWTaggedStringErrorCodeInvalidTags);
+            
+            // and without an error arg
+            string = [ZSWStringParser stringWithTaggedString:[ZSWTaggedString stringWithString:@"<tag><elephant>moo</tag></elephant>"]
+                                                     options:[ZSWTaggedStringOptions options]
+                                                 returnClass:[NSString class]
+                                                       error:nil];
+            expect(string).to.beNil();
         });
     });
     
@@ -40,12 +61,18 @@ describe(@"ZSWStringParser", ^{
         ZSWTaggedString *taggedString = [ZSWTaggedString stringWithString:@"no more timothy hay!"];
         
         NSArray *tags;
-        [[mockOptions expect] updateAttributedString:OCMOCK_ANY
-                                     updatedWithTags:[OCMArg capture:&tags]];
+        [[mockOptions expect] _private_updateAttributedString:OCMOCK_ANY
+                                              updatedWithTags:[OCMArg capture:&tags]];
         
-        expect([[ZSWStringParser stringWithTaggedString:taggedString
-                                                options:mockOptions
-                                            returnClass:[NSAttributedString class]] string]).to.equal(@"no more timothy hay!");
+        NSError *error;
+        
+        id string = [[ZSWStringParser stringWithTaggedString:taggedString
+                                                     options:mockOptions
+                                                 returnClass:[NSAttributedString class]
+                                                       error:&error] string];
+        
+        expect(string).to.equal(@"no more timothy hay!");
+        expect(error).to.beNil();
         expect(tags).to.haveCountOf(0);
     });
     
@@ -53,13 +80,18 @@ describe(@"ZSWStringParser", ^{
         ZSWTaggedString *taggedString = [ZSWTaggedString stringWithString:@"<all>circles presuppose</all>"];
         
         NSArray *tags;
-        [[mockOptions expect] updateAttributedString:OCMOCK_ANY
-                                     updatedWithTags:[OCMArg capture:&tags]];
+        [[mockOptions expect] _private_updateAttributedString:OCMOCK_ANY
+                                              updatedWithTags:[OCMArg capture:&tags]];
         
-        expect([[ZSWStringParser stringWithTaggedString:taggedString
-                                                options:mockOptions
-                                            returnClass:[NSAttributedString class]] string]).to.equal(@"circles presuppose");
+        NSError *error;
+        id string = [[ZSWStringParser stringWithTaggedString:taggedString
+                                                     options:mockOptions
+                                                 returnClass:[NSAttributedString class]
+                                                       error:&error] string];
+        
+        expect(string).to.equal(@"circles presuppose");
         expect(tags).to.haveCountOf(1);
+        expect(error).to.beNil();
         
         ZSWStringParserTag *tag = tags.firstObject;
         expect(tag.tagName).to.equal(@"all");
@@ -71,13 +103,18 @@ describe(@"ZSWStringParser", ^{
         ZSWTaggedString *taggedString = [ZSWTaggedString stringWithString:@"<some><all>circles presuppose</all></some>"];
         
         NSArray *tags;
-        [[mockOptions expect] updateAttributedString:OCMOCK_ANY
-                                     updatedWithTags:[OCMArg capture:&tags]];
+        [[mockOptions expect] _private_updateAttributedString:OCMOCK_ANY
+                                              updatedWithTags:[OCMArg capture:&tags]];
         
-        expect([[ZSWStringParser stringWithTaggedString:taggedString
-                                                options:mockOptions
-                                            returnClass:[NSAttributedString class]] string]).to.equal(@"circles presuppose");
+        NSError *error;
+        id string = [[ZSWStringParser stringWithTaggedString:taggedString
+                                                     options:mockOptions
+                                                 returnClass:[NSAttributedString class]
+                                                       error:&error] string];
+        
+        expect(string).to.equal(@"circles presuppose");
         expect(tags).to.haveCountOf(2);
+        expect(error).to.beNil();
         
         ZSWStringParserTag *someTag = tags[0];
         // some should go first, because outer ones are processed first
@@ -95,13 +132,18 @@ describe(@"ZSWStringParser", ^{
         ZSWTaggedString *taggedString = [ZSWTaggedString stringWithString:@"<all><qual>If I come <sad >without</sad> a thing</qual>, <suppose >then I come with <happy>all</happy> I need</suppose></all>"];
         
         NSArray *tags;
-        [[mockOptions expect] updateAttributedString:OCMOCK_ANY
-                                     updatedWithTags:[OCMArg capture:&tags]];
+        [[mockOptions expect] _private_updateAttributedString:OCMOCK_ANY
+                                              updatedWithTags:[OCMArg capture:&tags]];
         
-        expect([[ZSWStringParser stringWithTaggedString:taggedString
-                                                options:mockOptions
-                                            returnClass:[NSAttributedString class]] string]).to.equal(@"If I come without a thing, then I come with all I need");
+        NSError *error;
+        id string = [[ZSWStringParser stringWithTaggedString:taggedString
+                                                     options:mockOptions
+                                                 returnClass:[NSAttributedString class]
+                                                       error:&error] string];
+        
+        expect(string).to.equal(@"If I come without a thing, then I come with all I need");
         expect(tags).to.haveCountOf(5);
+        expect(error).to.beNil();
         
         ZSWStringParserTag *allTag = tags[0];
         expect(allTag.tagName).to.equal(@"all");
@@ -133,13 +175,18 @@ describe(@"ZSWStringParser", ^{
         ZSWTaggedString *taggedString = [ZSWTaggedString stringWithString:@"i am <hungry type=lol></hungry>banana"];
         
         NSArray *tags;
-        [[mockOptions expect] updateAttributedString:OCMOCK_ANY
-                                     updatedWithTags:[OCMArg capture:&tags]];
+        [[mockOptions expect] _private_updateAttributedString:OCMOCK_ANY
+                                              updatedWithTags:[OCMArg capture:&tags]];
         
-        expect([[ZSWStringParser stringWithTaggedString:taggedString
-                                                options:mockOptions
-                                            returnClass:[NSAttributedString class]] string]).to.equal(@"i am banana");
+        NSError *error;
+        id string = [[ZSWStringParser stringWithTaggedString:taggedString
+                                                     options:mockOptions
+                                                 returnClass:[NSAttributedString class]
+                                                       error:&error] string];
+        
+        expect(string).to.equal(@"i am banana");
         expect(tags).to.haveCountOf(1);
+        expect(error).to.beNil();
         
         ZSWStringParserTag *tag = tags.firstObject;
         expect(tag.tagName).to.equal(@"hungry");
@@ -153,13 +200,18 @@ describe(@"ZSWStringParser", ^{
         ZSWTaggedString *taggedString = [ZSWTaggedString stringWithFormat:@"i like to eat <eat>%@</eat>", ZSWEscapedStringForString(@"<apples>")];
         
         NSArray *tags;
-        [[mockOptions expect] updateAttributedString:OCMOCK_ANY
-                                     updatedWithTags:[OCMArg capture:&tags]];
+        [[mockOptions expect] _private_updateAttributedString:OCMOCK_ANY
+                                              updatedWithTags:[OCMArg capture:&tags]];
         
-        expect([[ZSWStringParser stringWithTaggedString:taggedString
-                                                options:mockOptions
-                                            returnClass:[NSAttributedString class]] string]).to.equal(@"i like to eat <apples>");
+        NSError *error;
+        id string = [[ZSWStringParser stringWithTaggedString:taggedString
+                                                     options:mockOptions
+                                                 returnClass:[NSAttributedString class]
+                                                       error:&error] string];
+        
+        expect(string).to.equal(@"i like to eat <apples>");
         expect(tags).to.haveCountOf(1);
+        expect(error).to.beNil();
         
         ZSWStringParserTag *tag = tags.firstObject;
         expect(tag.tagName).to.equal(@"eat");
@@ -171,12 +223,16 @@ describe(@"ZSWStringParser", ^{
         ZSWTaggedString *taggedString = [ZSWTaggedString stringWithFormat:@"i like to eat %@", ZSWEscapedStringForString(@"<")];
         
         NSArray *tags;
-        [[mockOptions expect] updateAttributedString:OCMOCK_ANY
-                                     updatedWithTags:[OCMArg capture:&tags]];
+        [[mockOptions expect] _private_updateAttributedString:OCMOCK_ANY
+                                              updatedWithTags:[OCMArg capture:&tags]];
         
-        expect([[ZSWStringParser stringWithTaggedString:taggedString
-                                                options:mockOptions
-                                            returnClass:[NSAttributedString class]] string]).to.equal(@"i like to eat <");
+        NSError *error;
+        id string = [[ZSWStringParser stringWithTaggedString:taggedString
+                                                     options:mockOptions
+                                                 returnClass:[NSAttributedString class]
+                                                       error:&error] string];
+        
+        expect(string).to.equal(@"i like to eat <");
         expect(tags).to.haveCountOf(0);
     });
 });